+++ src/Skeleton/index.tsx
@@ -1,7 +1,7 @@
 import React, { memo, useContext } from 'react';
 import styled, { keyframes, css } from 'styled-components/macro';
 import { ConfigContent } from '../ConfigProvider';
-import { pxToRem, pxTransform } from '../util';
+import { pxTransform } from '../util';
 import { ISkeletonProps } from './type';
 
 const skeletonLoading = keyframes`
@@ -16,6 +16,21 @@ const skeletonLoading = keyframes`
 const WrapperView = styled.div`
   width: 100%;
   display: flex;
+`;
+
+const AvatarView = styled.div`
+  width: ${({ base }: {base: number}) => pxTransform(48, base)};
+  height: ${({ base }: {base: number}) => pxTransform(48, base)};
+  background: linear-gradient(90deg, #f2f2f2 25%, #e6e6e6 37%, #f2f2f2 63%);
+  background-size: 400% 100%;
+  animation: ${skeletonLoading} 1.4s ease infinite;
+  border-radius: 50%;
+  flex: 0 0 auto;
+`;
+
+const ContentWrapView = styled.div`
+  padding-left: ${({ base }: {base: number}) => pxTransform(16, base)};
+  width: ${({ base }: {base: number}) => `calc(100% - ${pxTransform(48, base)})`};
 `;
 
 const ContentView = styled.div`
@@ -40,36 +55,40 @@ const TitleView = styled.h3`
 const ParagraphUlView = styled.ul`
   list-style: none;
   padding: 0;
-  margin: ${pxToRem(24)} 0 0;
+  margin: ${({ base }: {base: number}) => pxTransform(16, base)} 0 0;
 `;
 
 const ParagraphLiView = styled.li`
   width: 100%;
   ${ContentChildrenBaseStyle}
+  margin-top: ${({ base }: {base: number}) => pxTransform(8, base)};
 `;
 
 export default memo(({
-  title = true, rows = 3, loading = true, children,
+  title = true, avatar = false, rows = 3, loading = true, children, style = {},
 }: ISkeletonProps) => {
   // @ts-ignore
   const rowsArr = [...Array(rows)].map((i, index) => index);
   const rowsLen = rowsArr.length;
   const { baseFontSize } = useContext(ConfigContent);
-  const fontSizeProps = { base: baseFontSize };
+  const baseProps = { base: baseFontSize };
   return (
     <>
       {loading ? (
-        <WrapperView>
-          <ContentView>
-            {title && <TitleView {...fontSizeProps} />}
-            {rows > 0 && (
-              <ParagraphUlView>
-                {rowsArr.map((index) => (
-                  <ParagraphLiView {...fontSizeProps} key={index} style={{ width: index === rowsLen - 1 ? '60%' : '100%' }} />
-                ))}
-              </ParagraphUlView>
-            )}
-          </ContentView>
+        <WrapperView style={style}>
+          {avatar && <AvatarView {...baseProps} />}
+          <ContentWrapView {...baseProps} style={avatar ? {} : { paddingLeft: 0, width: '100%' }}>
+            <ContentView>
+              {title && <TitleView {...baseProps} />}
+              {rows > 0 && (
+                <ParagraphUlView {...baseProps}>
+                  {rowsArr.map((index) => (
+                    <ParagraphLiView {...baseProps} key={index} style={{ width: (rowsLen > 2 && index === rowsLen - 1) ? '60%' : '100%' }} />
+                  ))}
+                </ParagraphUlView>
+              )}
+            </ContentView>
+          </ContentWrapView>
         </WrapperView>
       ) : { children }}
     </>