+++ src/Skeleton/type.ts
@@ -1,7 +1,9 @@
-import { PropsWithChildren } from 'react';
+import { PropsWithChildren, CSSProperties } from 'react';
 
 export interface ISkeletonProps extends PropsWithChildren<any> {
+  style?: CSSProperties;
   rows?: number;
+  avatar?: boolean;
   title?: boolean;
   loading?: boolean;
 }