+++ src/Spin/index.spec.tsx
@@ -5,10 +5,8 @@ import 'jest-styled-components';
 
 describe('props', () => {
   it('test theme', () => {
-    const theme = {
-      color: '#2f54eb',
-    };
-    const tree = mount(<Spin {...{ theme }} />);
-    expect(tree.find(Line)).toHaveStyleRule('background-color', theme.color);
+    const color = '#2f54eb';
+    const tree = mount(<Spin {...{ color }} />);
+    expect(tree.find(Line)).toHaveStyleRule('background-color', color);
   });
 });