+++ src/TabBar/Tab.tsx
@@ -1,9 +1,11 @@
 import React, { memo } from 'react';
 import styled from 'styled-components/macro';
-import { pxToRem } from '../util';
+import { pxTransform, DEFAULT_BASE_FONT_SIZE } from '../util';
 
 export interface ITabProps {
+  key: string;
   selected?: boolean;
+  baseFontSize?: number;
   icon: string;
   selectedIcon: string;
   title: string;
@@ -21,22 +23,25 @@ const Wrapper = styled.div`
 `;
 
 const Title = styled.span`
-  margin-top: ${pxToRem(3)};
+  margin-top: ${({ base }: {base: number}) => pxTransform(3, base)};
   line-height: 1;
-  font-size: ${pxToRem(10)};
+  font-size: ${({ base }: {base: number}) => pxTransform(10, base)};
 `;
 
 const Icon = styled.div`
   background: transparent center / contain no-repeat;
-  width: ${pxToRem(22)};
-  height: ${pxToRem(22)};
+  width: ${({ base }: {base: number}) => pxTransform(22, base)};
+  height: ${({ base }: {base: number}) => pxTransform(22, base)};
 `;
 
 export default memo(({
-  selected = false, icon, selectedIcon, title, tintColor, unselectedTintColor, onPress,
-}: ITabProps) => (
-  <Wrapper style={{ color: selected ? tintColor : unselectedTintColor }} onClick={onPress}>
-    <Icon style={{ backgroundImage: `url(${selected ? selectedIcon : icon})` }} />
-    <Title>{title}</Title>
-  </Wrapper>
-));
+  selected = false, icon, selectedIcon, title, tintColor, unselectedTintColor, onPress, baseFontSize = DEFAULT_BASE_FONT_SIZE,
+}: ITabProps) => {
+  const baseProps = { base: baseFontSize };
+  return (
+    <Wrapper style={{ color: selected ? tintColor : unselectedTintColor }} onClick={onPress}>
+      <Icon {...baseProps} style={{ backgroundImage: `url(${selected ? selectedIcon : icon})` }} />
+      <Title {...baseProps}>{title}</Title>
+    </Wrapper>
+  );
+});